Moved tests for files into their own files
