+++ test/app.spec.js
@@ -1,29 +1,15 @@
 /* eslint-disable no-undef */
 /* eslint-disable no-unused-vars */
 
+console.log = () => { };
+
 const mocha = require('mocha');
 const request = require('supertest');
 const { expect } = require('chai');
+const Room = require('../room');
 const fs = require('fs');
 const { listener, rooms, socketJoinRoom } = require('../app');
-const Room = require('../room');
-const utils = require('../utils');
 
-describe('utils', function () {
-  it('should return local ip address', function (done) {
-    utils.getLocalIP().then(addr => {
-      expect(addr).to.contain('192.');
-      done();
-    }).catch(err => { done(err); });
-  });
-
-  it('should return a random room string of size 4', function (done) {
-    let str = utils.randomRoomString();
-    expect(str).to.contain('-');
-    expect(str.split('-')).to.be.lengthOf(4);
-    done();
-  });
-});
 
 describe('listener', function () {
   after(() => listener.close());
@@ -124,49 +110,4 @@ describe('socket', function () {
     }
     done();
   });
-});
-
-describe('room', function () {
-  it('should be empty on start', function (done) {
-    let room = new Room();
-    expect(room.isEmpty()).to.be.true;
-    done();
-  });
-
-  it('should contain two clients and not be empty', function (done) {
-    let room = new Room();
-    room.addClient({ id: 0, dummy: true });
-    room.addClient({ id: 1, dummy: true });
-    expect(room.numClients()).equals(2);
-    expect(room.isEmpty()).to.be.false;
-    done();
-  });
-
-  it('should show room is going to be deleted', function (done) {
-    after(() => clearTimeout(room.deleteTimer));
-    let room = new Room();
-    room.deleteTimer = setTimeout(() => { throw 'Should be cancelled'; }, 10000);
-    expect(room.willBeDeleted()).to.be.true;
-    done();
-  });
-
-  it('should show room will not be deleted anymore', function (done) {
-    let room = new Room();
-    room.deleteTimer = setTimeout(() => { throw 'Should be cancelled'; }, 10000);
-    room.cancelDeletion();
-    expect(room.willBeDeleted()).to.be.false;
-    expect(room.deleteTimer).to.be.null;
-    done();
-  });
-
-  it('should contain 2 then be empty', function (done) {
-    let room = new Room();
-    room.addClient({ id: 0, dummy: true });
-    room.addClient({ id: 1, dummy: true });
-    room.removeClient(0);
-    room.removeClient(1);
-    expect(room.isEmpty()).to.be.true;
-    done();
-  });
-
 });
